flask-otp-auth/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
argon2-cffi==23.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
twilio==8.11.1
aiohttp==3.9.1
redis==5.0.1
flask-limiter==3.5.0
email-validator>=2.1.0
requests>=2.31.0
